Replaced Layer with DDLayer's code. THE LAYER IS DEAD! LONG LIVE THE LAYER!
