+++ src/main/java/C7/Model/Layer/Layer.java
@@ -8,15 +8,20 @@ import java.util.*;
 /**
  * Represents a basic image layer.
  * @author Elias Ersson
+ * @author Isak Gustafsson
  * @author Love Gustafsson
  * @author Hugo Ekstrand
  * @version 2.2
  */
-class Layer implements ILayer, Serializable {
+public class Layer implements ILayer, Serializable {
 
     private transient Collection<IObserver<Tuple2<Vector2D, Vector2D>>> observers = new ArrayList<>();
 
-    private Color[][] pixels;   // This layer's pixel data.
+    private float[][] r;
+    private float[][] g;
+    private float[][] b;
+    private float[][] a;
+
     private int width;          // The width, in pixels, of this layer.
     private int height;         // The height, in pixels of this layer.
 
@@ -41,14 +46,14 @@ class Layer implements ILayer, Serializable {
         if(width < 0 || height < 0)
             throw new IllegalArgumentException();
 
-        pixels = new Color[width][height];
+        setRBGASize(width,height);
         this.width  = width;
         this.height = height;
 
         for (int x = 0; x < width; x++) {
             for (int y = 0; y < height; y++) {
                 // Copy color values
-                pixels[x][y] = new Color(color);
+                setColor(x,y, color);
             }
         }
     }
@@ -58,19 +63,15 @@ class Layer implements ILayer, Serializable {
      * @param colorMatrix The matrix of color data of the new layer
      */
     public Layer(Color[][] colorMatrix) {
-        /*pixels = colorMatrix; OPTIMIZATION this works faster but is technically bad because
-        * it leaves an open reference somewhere to the new layer's color data
-        * DEFENSIVE COPYING */
-
-        pixels = new Color[colorMatrix.length][];
+        //Avoids crash if width = 0
+        if (colorMatrix.length == 0) setRBGASize(colorMatrix.length, 0);
+        else setRBGASize(colorMatrix.length, colorMatrix[0].length);
 
         for (int x = 0; x < colorMatrix.length; x++) {
-            pixels[x] = new Color[colorMatrix[x].length];
-
             for (int y = 0; y < colorMatrix[x].length; y++) {
                 Color oldColor = colorMatrix[x][y];
                 //Would like a copy method for copying color data instead of this repeating pattern of GET
-                pixels[x][y] = new Color(oldColor.getRed(), oldColor.getGreen(), oldColor.getBlue(), oldColor.getAlpha());
+                setColor(x, y, oldColor.getRed(), oldColor.getGreen(), oldColor.getBlue(), oldColor.getAlpha());
             }
         }
 
@@ -78,6 +79,36 @@ class Layer implements ILayer, Serializable {
         width = colorMatrix.length;
         if (width == 0) height = 0;
         else height = colorMatrix[0].length;
+    }
+
+    //Color data modification methods
+    //Set layer size
+    private void setRBGASize(int width, int height){
+        r = new float[width][height];
+        g = new float[width][height];
+        b = new float[width][height];
+        a = new float[width][height];
+
+        this.width = width;
+        this.height = height;
+    }
+    //set the color of a pixel
+    private void setColor(int x, int y, float r, float g, float b, float a){
+        this.r[x][y] = r;
+        this.g[x][y] = g;
+        this.b[x][y] = b;
+        this.a[x][y] = a;
+    }
+    //set the color of a pixel with a Color
+    private void setColor(int x, int y, Color color){
+        r[x][y] = color.getRed();
+        g[x][y] = color.getGreen();
+        b[x][y] = color.getBlue();
+        a[x][y] = color.getAlpha();
+    }
+    //create Color from the color of a pixel
+    private Color getColor(int x, int y){
+        return new Color(r[x][y], g[x][y], b[x][y], a[x][y]);
     }
 
     @Override
@@ -93,8 +124,8 @@ class Layer implements ILayer, Serializable {
     @Override
     public Color getLocalPixel(int x, int y) {
         if(!isPixelOnLocalLayer(x, y))
-           throw new IllegalArgumentException();
-        return pixels[x][y];
+            throw new IllegalArgumentException();
+        return getColor(x,y);
     }
 
     @Override
@@ -113,7 +144,7 @@ class Layer implements ILayer, Serializable {
 
         Vector2D global = toGlobal(new Vector2D(x, y));
         updateRectangleOfChange((int)global.getX(), (int)global.getY());
-        pixels[x][y] = color;
+        setColor(x,y, color);
     }
 
     /**
@@ -267,13 +298,25 @@ class Layer implements ILayer, Serializable {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Layer layer = (Layer) o;
-        return width == layer.width && height == layer.height && Arrays.deepEquals(pixels, layer.pixels);
+        if(width != layer.width || height != layer.height) return false;
+        for (int x = 0; x < r.length; x++) {
+            for (int y = 0; y < r[0].length; y++) {
+                Color layerColor = layer.getLocalPixel(x, y);
+                if (
+                        r[x][y] != layerColor.getRed() ||
+                                g[x][y] != layerColor.getGreen() ||
+                                b[x][y] != layerColor.getBlue() ||
+                                a[x][y] != layerColor.getAlpha()
+                ) return false;
+            }
+        }
+        return true;
     }
 
     @Override
     public int hashCode() {
         int result = Objects.hash(width, height);
-        result = 31 * result + Arrays.hashCode(pixels);
+        result = 31 * result + Arrays.hashCode(r) + Arrays.hashCode(g) + Arrays.hashCode(b) + Arrays.hashCode(b);
         return result;
     }
 
@@ -294,24 +337,30 @@ class Layer implements ILayer, Serializable {
         if(width < 0 || height < 0)
             throw new IllegalArgumentException();
 
-        Color[][] newPixels = new Color[width][height];
+        float[][]oldR = r;
+        float[][]oldG = g;
+        float[][]oldB = b;
+        float[][]oldA = a;
+        int oldW = this.width;
+        int oldH = this.height;
+
+        setRBGASize(width, height);
 
         final Color emptyColor = new Color(0, 0, 0, 1);
 
         for (int x = 0; x < width; x++) {
-            if(x < this.width){
-                System.arraycopy(pixels[x], 0, newPixels[x], 0, Math.min(pixels[x].length, newPixels[x].length) - 1);
-                if(newPixels[x].length - pixels[x].length > 0)
-                    Arrays.fill(newPixels[x], pixels[x].length, newPixels[x].length, emptyColor);
-            }
-            else{
-                Arrays.fill(newPixels[x], emptyColor);
+            for (int y = 0; y < height; y++) {
+                //Exists old color
+                if (oldW - x > 0 && oldH - y > 0){
+                    setColor(x, y, oldR[x][y], oldG[x][y], oldB[x][y], oldA[x][y]);
+                }
+                //No old color just new
+                else{
+                    setColor(x, y, emptyColor);
+                }
             }
         }
 
-        pixels      = newPixels;
-        this.width  = width;
-        this.height = height;
         maxRectangleOfChange();
     }
 
@@ -348,7 +397,7 @@ class Layer implements ILayer, Serializable {
         return this;
     }
 
-    @java.lang.Override
+    @Override
     public LayerFormat getFormat() {
         return LayerFormat.RGBA32F32F32F32F;
     }
